Added stats information
