+++ api/crime/crime_schema.js
@@ -17,22 +17,4 @@ type CrimeIncident {
   offense_group_short_description: String
   offense_group_long_description: String
 }
-type CrimeIncidentStats {
-  incidents: [CrimeIncident]
-  date: String
-  count: Int
-  agency: String
-  date_occurred: String
-  month: Int
-  year: Int
-}
-type GenericMonthStats {
-  count: Int
-  month: Int
-  year: Int
-  day: Int
-  grouptitle: String,
-  groupcategory: String,
-  subitems: [GenericMonthStats]
-}
 `;
