+++ api/simplicity_schema.js
@@ -60,6 +60,7 @@ const streetSchema = require('./street/street_schema');
 const neighborhoodSchema = require('./neighborhood/neighborhood_schema');
 const cipSchema = require('./cip/cip_schema');
 const financeSchema = require('./finance/finance_schema');
+const statsSchema = require('./crime/stats_schema');
 
 const schema = [
   baseSchema,
